+++ zarrs_plugin/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Use `paste` re-export in the `impl_extension_aliases` macro
+
 ## [0.3.0] - 2025-12-31
 
 ### Added